requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyascii==0.3.2
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
bangla==0.0.2
bcrypt==4.1.2
blinker==1.7.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dateparser==1.1.8
decorator==5.1.1
dnspython==2.6.1
docopt==0.6.2
einops==0.7.0
encodec==0.1.1
ffmpeg-python==0.2.0
filelock==3.13.4
Flask==3.0.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
g2pkk==0.1.2
grpcio==1.62.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.0
gruut_lang_en==2.0.0
gruut_lang_es==2.0.0
gruut_lang_fr==2.0.2
hangul-romanize==0.1.0
huggingface-hub==0.22.2
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflect==7.2.0
itsdangerous==2.1.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.4.0
jsonlines==1.2.0
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.42.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
more-itertools==10.2.0
motor==3.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.59.1
numpy==1.22.0
packaging==24.0
pandas==1.5.3
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
preshed==3.0.9
protobuf==5.26.1
psutil==5.9.8
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
pynndescent==0.5.12
pyparsing==3.1.2
pypinyin==0.51.0
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thinc==8.2.3
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.2.2
torchaudio==2.2.2
tornado==6.4
tqdm==4.66.2
trainer==0.0.36
transformers==4.39.3
TTS==0.22.0
typeguard==4.2.1
typer==0.9.4
typing_extensions==4.10.0
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.2
whispercpp @ git+https://github.com/stlukey/whispercpp.py@7af678159c29edb3bc2a51a72665073d58f2352f
yarl==1.9.4
zipp==3.18.1
